create loops dir; arrange manual_flatten lint and utils
